Remove ACL from PermittedContract, and create AclPermittedContract
