+++ SmartContracts.NET/ContractsCore/Contracts/PermittedContract.cs
@@ -8,24 +8,12 @@ namespace ContractsCore.Contracts
 {
 	public abstract class PermittedContract : Contract
 	{
-		private readonly AccessControlList acl;
-
-		protected PermittedContract(Address address, Address permissionManager)
-			: this(address, permissionManager, new AccessControlList())
-		{
-		}
-
-		protected PermittedContract(Address address, Address permissionManager, AccessControlList acl)
+		protected PermittedContract(Address address)
 			: base(address)
 		{
-			this.acl = acl;
-			this.ConfigurePermissionManager(permissionManager);
 		}
 
-		public bool HasPermission(Address address, Permission permission)
-		{
-			return this.acl.HasPermission(address, permission);
-		}
+		public abstract bool HasPermission(Address address, Permission permission);
 
 		protected internal override bool Receive(Action action)
 		{
@@ -52,28 +40,14 @@ namespace ContractsCore.Contracts
 		protected void RequirePermission(Action action)
 		{
 			var permission = new Permission(action.GetType());
-			if (!this.acl.HasPermission(action.Sender, permission))
+			if (!this.HasPermission(action.Sender, permission))
 			{
 				throw new NoPermissionException(this, action.Sender, permission);
 			}
 		}
 
-		private void ConfigurePermissionManager(Address permissionManager)
-		{
-			this.acl.AddPermission(permissionManager, new Permission(typeof(AddPermissionAction)));
-			this.acl.AddPermission(permissionManager, new Permission(typeof(RemovePermissionAction)));
-		}
+		protected abstract void HandleAddPermissionAction(AddPermissionAction action);
 
-		private void HandleAddPermissionAction(AddPermissionAction action)
-		{
-			this.RequirePermission(action);
-			this.acl.AddPermission(action.PermittedAddress, action.Permission);
-		}
-
-		private void HandleRemovePermissionAction(RemovePermissionAction action)
-		{
-			this.RequirePermission(action);
-			this.acl.RemovePermission(action.PermittedAddress, action.Permission);
-		}
+		protected abstract void HandleRemovePermissionAction(RemovePermissionAction action);
 	}
 }
